overlay: fix inappropriate data type
Problem: parent->lastsent is stored as an int but holds the result
of flux_reactor_now(), which returns a double.

This was called out by github-advanced-security in CI.

Change lastsent from int to double.

+++ src/modules/overlay/parent.h
@@ -26,7 +26,7 @@ struct parent {
     void *zsock;            // NULL on rank 0
     char *uri;
     flux_watcher_t *w;
-    int lastsent;
+    double lastsent;
     char *pubkey;
     uint32_t rank;
     char uuid[UUID_STR_LEN];
